requirements.txt
Flask==3.0.3
numpy==1.26.4
opencv-python==4.10.0.84
scipy==1.13.1
python-dotenv==1.0.1
google-generativeai==0.8.3
gunicorn==22.0.0
